wm_server/requirements.txt
# Runtime dependencies for WorkspaceManager server (wm_server)
# These are minimal runtime-only requirements. Test/dev tools are intentionally excluded.

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
docker>=7.0.0
python-multipart>=0.0.9
pytest>=7.4.0
python-dotenv>=1.0.0

